qa(tests): Implement tests that expect an exception
WIP - experimental attempt at fixing the remoteContexts issue. NOT WORKING (for now)

WIP - add more fixes for tests with errors. It broke a test that was fine however

architecture(tests): refactor the exceptions and the tests exceptions

WIP - fix more exception tests. One previsouly working test is now broken however

WIP on tests with exceptions : fix the issue with `isIri()` being too permissive and replace some calls to this method with `isAbsoluteIri()`, which is not in the specs but seem better

WIP on exceptions tests : they all pass! Only need to do so refactoring and we are good

qa(tests): Implement tests that expect an exception